requirements.txt
cupy-cuda11x>=11.0.0
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
jupyter>=1.0.0
pytest>=7.0.0
pandas>=1.3.0
scikit-image>=0.19.0
Pillow>=8.3.0
opencv-python>=4.5.0
